+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 
 ##### Enhancements
 
+* Add `quick_discouraged_focused_test` opt-in rule which warns against
+  focused tests in Quick tests.  
+  [Ornithologist Coder](https://github.com/ornithocoder)
+  [#1905](https://github.com/realm/SwiftLint/issues/1905)
+
 * Add `override_in_extension` opt-in rule that warns against overriding
   declarations in an `extension`.  
   [Marcelo Fabri](https://github.com/marcelofabri)